server/requirements.txt
flask>=3.1.0
werkzeug>=2.3.0
flask-socketio==5.3.3
flask-cors==3.0.10
python-dotenv==1.0.0
anthropic>=0.18.1
SpeechRecognition==3.10.0
eventlet==0.33.3
gunicorn==20.1.0